// src/types/trendyContentInterface.ts
// Generated by https://quicktype.io

export interface TrendyContent {
  page: number;
  results: TrendyContentResult[];
  total_pages: number;
  total_results: number;
}

export interface TrendyContentResult {
  adult: boolean;
  backdrop_path: string;
  id: number;
  title?: string;
  original_language: OriginalLanguage;
  original_title?: string;
  overview: string;
  poster_path: string;
  media_type: MediaType;
  genre_ids: number[];
  popularity: number;
  release_date?: string;
  video?: boolean;
  vote_average: number;
  vote_count: number;
  name?: string;
  original_name?: string;
  first_air_date?: string;
  origin_country?: string[];
}

export enum MediaType {
  Movie = 'movie',
  Tv = 'tv',
}

export enum OriginalLanguage {
  De = 'de',
  En = 'en',
  Fr = 'fr',
}